Refactor PWLocal
